# Use entire road geometry for NameSimilarityEdgeFilter