Fixes #2319