@easbar 2021-05-16
I had to fix the routing measurement for point_hint, because so far we always set the point_hint to the name of the very first edge. But to test the speed of the look-ups with point hints I think we should rather use some point hint that cannot be found (worst case scenario in terms of performance). I fixed this here: b8a25d775b4578b9e37912736cfbccdd49ea9f25 and it changed the measured routing CH speed with point hints from about 1.5ms to 1.8ms, so we can probably say that point_hints make route calculation (including location lookup, but no instructions, serialization etc.) about 20% slower.

After fixing this I compared the performance (`metrics.routingCH_with_hints.mean`) of this PR vs. master:

|map|master|this branch|
|-|-|-|
|germany|1.78ms|1.83ms|
|bayern|1.19ms|1.21ms|

The difference is less than 3% even on the smaller map where I assume the relative amount of time spent to lookup the locations is larger (because the routes are shorter). I think we can conclude that the bbox/geometry check does not play an important role and not too surprisingly the majority of the additional time we need to use `point_hint` is spent on the string comparison checks.

Btw I also did a quick performance comparison for the current scenario where we only use the tower nodes. Not using a `PointList` via `fetchWayGeometry` and instead accessing the tower node coordinates directly like `nodeAccess.getLat(iter.getBaseNode())` did not make a significant difference here either, <2%.